requirements.txt
Django==2.1.2
django-markdownx==2.0.23
Markdown==2.6.11
Pillow==5.1.0
pytz==2018.4
